Include polygons at node level
update the drift code to ingest the polygon that comes out of the aerial toolbox if it is available. This is passed as a user argument, so the toolbox operator decides if they want to pass a path to a folder containing a relevant shapefile. This will then find all nodes present there and apply them while leaving the rest as bboxes